requirements.txt

langgraph==1.0.3
python-dotenv==1.1.1
ddgs==9.5.2
markitdown[all]==0.1.2
fastapi==0.111.0
uvicorn[standard]==0.38.0
langchain-groq==1.0.1
pydantic-settings==2.3.0
langchain==0.2.7
langchain-core==0.2.7
